+++ internal/promapi/config.go
@@ -42,7 +42,7 @@ func (q configQuery) Run() queryResult {
 		Str("uri", q.prom.safeURI).
 		Msg("Getting prometheus configuration")
 
-	ctx, cancel := context.WithTimeout(q.ctx, q.prom.timeout)
+	ctx, cancel := q.prom.requestContext(q.ctx)
 	defer cancel()
 
 	var qr queryResult